"Feat : 통합 테스트 코드 예시 작성"
테스트 코드 골격 계획
DB 확인하는 방법과 응답 내용 확인하는 방법 기입
