+++ .gitignore
@@ -38,3 +38,4 @@ out/
 
 # Project Gitignore
 src/main/resources/application-local.properties
+/src/test/resources/.env
