+++ src/test/java/com/example/demo/utils/LoadEnvironmentVariables.java
@@ -1,0 +1,14 @@
+package com.example.demo.utils;
+
+import org.springframework.test.context.TestPropertySource;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@TestPropertySource(locations = "classpath:.env")
+public @interface LoadEnvironmentVariables {
+}
