+++ src/test/resources/application-test.properties
@@ -1,0 +1,47 @@
+spring.datasource.url=jdbc:h2:mem:db
+spring.datasource.username=${H2_USERNAME}
+spring.datasource.password=${H2_PASSWORD}
+spring.datasource.driver-class-name=org.h2.Driver
+
+spring.jpa.hibernate.ddl-auto=create
+
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.properties.hibernate.use_sql_comments=true
+
+jwt.secret.key=${JWT_SECRET_KEY}
+
+#S3
+cloud.aws.credentials.accessKey=${S3_ACCESS_KEY}
+cloud.aws.credentials.secretKey=${S3_SECRET_KEY}
+cloud.aws.region.static=${S3_REGION_STATIC}
+cloud.aws.stack.auto=false
+cloud.aws.s3.bucket=${S3_BUCKET_NAME}
+
+#Swagger
+springdoc.packages-to-scan=com.example.demo
+springdoc.default-consumes-media-type=application/json;charset=UTF-8
+springdoc.default-produces-media-type=application/json;charset=UTF-8
+springdoc.swagger-ui.path=/112
+springdoc.swagger-ui.disable-swagger-default-url=true
+springdoc.swagger-ui.display-request-duration=true
+springdoc.swagger-ui.operations-sorter=alpha
+springdoc.swagger-ui.enabled=true
+
+#H2 DB Console
+spring.h2.console.enabled=true
+spring.h2.console.path=/console
+spring.h2.console.settings.web-allow-others=true
+
+#TestCase in DB
+spring.sql.init.data-locations=classpath:data/data-test.sql
+spring.jpa.defer-datasource-initialization=true
+
+#OAuth2KaKao
+auth.oauth2.kakao.client-id=851d6c649ed19d32afa2743c91134e57
+auth.oauth2.kakao.redirect-url=http://localhost:5173/api/auth/kakao/callback
+
+#redis
+spring.data.redis.host=${REDIS_HOST}
+spring.data.redis.password=${REDIS_PASSWORD}
+spring.data.redis.port=14859
