+++ src/test/resources/application.properties
@@ -1,0 +1,4 @@
+spring.profiles.active=${PROFILE}
+spring.devtools.livereload.enabled=true
+spring.devtools.restart.enabled=false
+spring.freemarker.cache=false
