+++ src/test/resources/data/data-test.sql
@@ -1,0 +1,2 @@
+-- 오류 방지를 위한 아무런 의미 없는 DML
+select 1;
