+++ src/test/resources/testcase-member.sql
@@ -1,0 +1,17 @@
+insert into member (nickname, username, image) values ('hwassell0', 1, 'http://dummyimage.com/216x202.png/dddddd/000000');
+insert into member (nickname, username, image) values ('dblasoni1', 2, 'http://dummyimage.com/236x229.png/dddddd/000000');
+insert into member (nickname, username, image) values ('nleigh2', 3, 'http://dummyimage.com/176x176.png/cc0000/ffffff');
+insert into member (nickname, username, image) values ('fabramow3', 4, 'http://dummyimage.com/208x209.png/cc0000/ffffff');
+insert into member (nickname, username, image) values ('dtegler4', 5, 'http://dummyimage.com/166x235.png/cc0000/ffffff');
+
+insert into member_location (member_id, name, latitude, longitude) values (1, '7736 Lighthouse Bay Pass', '531', '487');
+insert into member_location (member_id, name, latitude, longitude) values (2, '77647 Chive Pass', '699', '540');
+insert into member_location (member_id, name, latitude, longitude) values (3, '5 Shoshone Avenue', '234', '455');
+insert into member_location (member_id, name, latitude, longitude) values (4, '6 Garrison Lane', '651', '516');
+insert into member_location (member_id, name, latitude, longitude) values (5, '45 Pine View Hill', '503', '447');
+
+insert into shop (seller_id, intro, name) values (1, 'Etiam vel augue. Vestibulum rutrum rutrum neque. Aenean auctor gravida sem.', 'jbiggen0');
+insert into shop (seller_id, intro, name) values (2, 'Ut at dolor quis odio consequat varius. Integer ac leo.', 'cmaccurlye1');
+insert into shop (seller_id, intro, name) values (3, 'Aliquam quis turpis eget elit sodales scelerisque.', 'santonin2');
+insert into shop (seller_id, intro, name) values (4, 'Duis ac nibh. Fusce lacus purus, aliquet at, feugiat non, pretium quis, lectus.', 'cfrounks3');
+insert into shop (seller_id, intro, name) values (5, 'Duis ac nibh.', 'caylett4');
